+++ lib/core/src/ng-progress.module.ts
@@ -1,22 +1,31 @@
-import { NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { NgProgressBarComponent } from './components/ng-progress-bar/ng-progress-bar.component';
-import { NgProgressComponent } from './components/ng-progress/ng-progress.component';
-import { NgProgress } from './services/ng-progress.service';
+import { NgProgressComponent } from './ng-progress.component';
+import { NgProgress } from './ng-progress.service';
+import { NgProgressConfig } from './ng-progress.interface';
+import { CONFIG } from './ng-progress.token';
+
+export function NgProgressFactory(config: NgProgressConfig) {
+  return new NgProgress(config);
+}
 
 @NgModule({
-  declarations: [
-    NgProgressComponent,
-    NgProgressBarComponent
-  ],
+  declarations: [NgProgressComponent],
   exports: [NgProgressComponent],
   imports: [CommonModule]
 })
 export class NgProgressModule {
-  static forRoot() {
+  static forRoot(config?: NgProgressConfig): ModuleWithProviders {
     return {
       ngModule: NgProgressModule,
-      providers: [NgProgress]
+      providers: [
+        {provide: CONFIG, useValue: config},
+        {
+          provide: NgProgress,
+          useFactory: NgProgressFactory,
+          deps: [CONFIG]
+        }
+      ]
     };
   }
 }