+++ lib/router/src/index.ts
@@ -1,1 +1,1 @@
-export * from './ng-progress-router.module'
+export * from './ng-progress-router.module';