+++ src/app/header/header.component.ts
@@ -4,7 +4,8 @@ import {Component, ChangeDetectionStrategy} from '@angular/core';
   selector: 'app-header',
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  preserveWhitespaces: false
 })
 export class HeaderComponent {